requirements.txt
# Core dependencies
requests>=2.28.0
feedparser>=6.0.0
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Whisper options (install one)
# Option 1: faster-whisper (recommended - 4x faster)
faster-whisper>=0.10.0

# Option 2: openai-whisper (fallback)
# openai-whisper>=20231117

# AI Summarization (install one or both)
anthropic>=0.18.0
openai>=1.12.0

# HTTP API Server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
aiofiles>=23.0.0

# GCP deployment
google-cloud-storage>=2.10.0
google-cloud-speech>=2.20.0
